Extract thread automation storage
